+++ doc/changelog.md
@@ -3,6 +3,7 @@
 * **0.15.8** (2016-06-25)
   - Removed image configuration caching (#495)
   - Fix for tcp wait when used with Docker for Mac (#430)
+  - Add warning when assembly is empty when watching a Docker image (#490)
   
 * **0.15.7** (2016-06-09)
   - Add support for '.maven-dockerinclude' for including certain files in plain Dockerfile build ([#471](https://github.com/fabric8io/docker-maven-plugin/issues/471))